requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
distro==1.8.0
Django==4.2.2
django-debug-toolbar==4.2.0
django-pytest==0.2.0
django-redis==5.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
fontawesomefree==6.3.0
idna==3.4
iniconfig==2.0.0
kombu==5.3.1
netaddr==0.8.0
packaging==23.1
Pillow==10.0.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2==2.9.6
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.4
var-dump==1.2
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
yookassa==2.4.0
